Added Int and Word (except Word64) to Arg. Refs #10.
Word64 should be added later. Writing unsigned 64 bit literals in C requires
using a L suffix (or other sizes, depending on the architecture). Copilot does
not handle this correctly right now. We need to find a way to make this work
for every architecture.

For now, we leave these large numbers out of the tests.