Disabled the orphans warnings in GHC. Refs #10.
We can't work around the orphaned issue: Arbitrary is defined in quickcheck,
the datatypes are defined in different files of copilot-language. If we want to
keep them optional, the best approach is the have the current situation (where
the instances are in a seperate directory). Hiding these warnings makes it
easier to see the ones we are interested in.